requirements.txt
Django==4.2.7
django-admin-sortable2==1.0.4